[MISC] Rename seqan3::views::type_reduce_view
