+++ include/seqan3/utility/views/interleave.hpp
@@ -19,11 +19,11 @@
 #include <seqan3/core/detail/persist_view.hpp>
 #include <seqan3/core/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 #include <seqan3/utility/type_traits/pre.hpp>
 #include <seqan3/utility/views/chunk.hpp>
 #include <seqan3/utility/views/join_with.hpp>
+#include <seqan3/utility/views/type_reduce.hpp>
 
 namespace seqan3::detail
 {
