+++ test/unit/utility/views/CMakeLists.txt
@@ -8,4 +8,5 @@ seqan3_test(repeat_n_test.cpp)
 seqan3_test(repeat_test.cpp)
 seqan3_test(single_pass_input_test.cpp)
 seqan3_test(slice_test.cpp)
+seqan3_test(type_reduce_test.cpp)
 seqan3_test(zip_test.cpp)
